# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.TriggerSyncRequest:
    properties:
      ad_account_id:
        example: act_1234567890
        type: string
      date_start:
        example: "2026-03-01"
        type: string
      date_stop:
        example: "2026-03-31"
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      errors: {}
      message:
        example: Error message
        type: string
      success:
        example: false
        type: boolean
    type: object
  response.Meta:
    properties:
      last_page:
        type: integer
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  response.MetaPaging:
    properties:
      cursors:
        properties:
          after:
            type: string
          before:
            type: string
        type: object
      has_next:
        type: boolean
      has_previous:
        type: boolean
    type: object
...
  response.PaginationResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        example: true
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      paging:
        $ref: '#/definitions/response.MetaPaging'
      success:
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      success:
        example: true
        type: boolean
    type: object
  sync.MetaSyncBatch:
    properties:
      ad_account_id:
        type: string
      ad_account_name:
        type: string
      batch_code:
        type: string
      created_at:
        type: string
      date_preset:
        type: string
      date_start:
        type: string
      date_stop:
        type: string
      duration_ms:
        type: integer
      error_message:
        type: string
      failed_count:
        type: integer
      finished_at:
        type: string
      id:
        type: integer
      inserted_count:
        type: integer
      metadata:
        type: object
      provider:
        type: string
      rate_limit_hit:
        type: boolean
      request_count:
        type: integer
      skipped_count:
        type: integer
      started_at:
        type: string
      status:
        type: string
      steps:
        items:
          $ref: '#/definitions/sync.MetaSyncStep'
        type: array
      sync_mode:
        type: string
      sync_scope:
        type: string
      total_records:
        type: integer
      updated_at:
        type: string
      updated_count:
        type: integer
    type: object
  sync.MetaSyncStep:
    properties:
      batch:
        $ref: '#/definitions/sync.MetaSyncBatch'
      batch_id:
        type: integer
      created_at:
        type: string
      cursor_after:
        type: string
      cursor_before:
        type: string
      duration_ms:
        type: integer
      endpoint:
        type: string
      error_code:
        type: string
      error_message:
        type: string
      failed_count:
        type: integer
      finished_at:
        type: string
      has_next:
        type: boolean
      id:
        type: integer
      inserted_count:
        type: integer
      metadata:
        type: object
      request_count:
        type: integer
      skipped_count:
        type: integer
      started_at:
        type: string
      status:
        type: string
      sync_type:
        type: string
      total_records:
        type: integer
      updated_at:
        type: string
      updated_count:
        type: integer
    type: object
host: localhost:8888
info:
  contact: {}
  description: ERP Ready Backend for Meta Ads & Business Management
  title: Ads Backend API
  version: "1.0"
paths:
...
  /meta/sync:
    post:
      description: Manually trigger a full Meta Ads sync. Returns immediately; subscribe
        to the Centrifugo channel for real-time progress.
      parameters:
      - description: Sync Request (optional ad_account_id, date_start, date_stop)
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.TriggerSyncRequest'
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Sync already in progress
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Trigger Meta Ads Sync
      tags:
      - Meta Sync
  /meta/sync/batches:
    get:
      description: Returns a paginated list of sync batches ordered by most recent
        first.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 25
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/sync.MetaSyncBatch'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Sync Batches
      tags:
      - Meta Sync
  /meta/sync/batches/{id}:
    get:
      description: Returns a single sync batch with its steps.
      parameters:
      - description: Batch ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/sync.MetaSyncBatch'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Sync Batch
      tags:
      - Meta Sync
  /meta/sync/status:
    get:
      description: Returns whether a sync job is currently running.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Sync Status
      tags:
      - Meta Sync
...
securityDefinitions:
  BearerAuth:
    description: Tuliskan 'Bearer ' diikuti dengan token JWT Anda.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"